Migrate Selection tests to munit
